CIW
I went ahead and made the application project on my own for this chapter despite not needing to. I figured I should make you aware of that since there are a couple of extra files added to this commit because of that. Lab 9-3 and the Optional Lab did not need any changes to be made in the editor.